[add/update] refs #58 ディレクトリの階層を整理
 保守性改善のため
