+++ app/_common/constants/Colors.ts
@@ -1,1 +1,1 @@
-export const Colors = ['#0088FE', '#00C49F', '#FFBB28', '#FF8042']
+export const colors = ['#0088FE', '#00C49F', '#FFBB28', '#FF8042']
