+++ app/_common/types/Account.ts
@@ -1,0 +1,5 @@
+export type Account = {
+  id: number
+  name: string
+  usage: string
+}
