+++ app/_common/types/AssetParFinancialInstitution.ts
@@ -1,0 +1,10 @@
+export type AssetParFinancialInstitution = {
+  name: string | null
+  usage: string | null
+  Asset: ParFinancialInstitutionAsset[]
+}
+
+export type ParFinancialInstitutionAsset = {
+  date: string | null
+  amount: number | null
+}
