+++ app/_common/types/FinancialInstitution.ts
@@ -1,0 +1,4 @@
+export type FinancialInstitution = {
+  name: string | null
+  usage: string | null
+}
