+++ app/_components/finance/parts/finance-form.tsx
@@ -1,11 +1,12 @@
 'use client'
+
 import { XMarkIcon } from '@heroicons/react/24/solid'
-import { Card, CardBody, Input, Button, Typography } from '../common'
+import { Card, CardBody, Input, Button, Typography } from '../../common'
 import { Dispatch, SetStateAction, useState } from 'react'
 import { z } from 'zod'
-import Loading from '../../(routes)/loading'
+import Loading from '../../../(routes)/loading'
 import { createClientComponentClient } from '@supabase/auth-helpers-nextjs'
-import { Database } from '../../../lib/database.types'
+import { Database } from '../../../../lib/database.types'
 import { SubmitHandler, useForm } from 'react-hook-form'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { useRouter } from 'next/navigation'
@@ -13,28 +14,43 @@ type Schema = z.infer<typeof schema>
 
 // 入力データの検証ルールを定義
 const schema = z.object({
-  financeInstitution: z.string().min(1, {
-    message: '金融機関を入力してください。',
+  financeInstitution: z.string({
+    required_error: '金融機関を入力してください。',
   }),
-  usage: z.string().min(1, {
-    message: '使い方を入力してください。',
+  date: z
+    .date({
+      required_error: '日付を入力してください。',
+      invalid_type_error: "日付は'yyyy/mm/dd'の形式で入力してください。",
+    })
+    .transform((val) => val.toISOString().slice(0, 10)),
+  amount: z.number({
+    required_error: '金額を入力してください。',
+    invalid_type_error: '金額は数値で入力してください。',
   }),
 })
 
-export default function AcountAddForm({
-  open,
-  setOpen,
-  id,
-}: {
+type Props = {
   open: boolean
   setOpen: Dispatch<SetStateAction<boolean>>
-  id: string
-}) {
+  name: string
+  id: string | undefined
+}
+
+/**
+ * 金額入力フォーム
+ * @param open モーダルウィンドウの表示状態
+ * @param setOpen モーダルウィンドウの表示状態を変更する関数
+ * @param name 金融機関
+ * @param id ユーザーID
+ */
+export default function PopUpForm({ open, setOpen, name, id }: Props) {
   const router = useRouter()
   const supabase = createClientComponentClient<Database>()
   const [loading, setLoading] = useState(false)
   const [message, setMessage] = useState('')
 
+  // 現在の日付を取得
+  const today = new Date().toISOString().slice(0, 10)
   // 入力フォームの設定
   const {
     register,
@@ -43,7 +59,7 @@ export default function AcountAddForm({
     reset,
   } = useForm({
     // 初期値
-    defaultValues: { financeInstitution: '', usage: '' },
+    defaultValues: { financeInstitution: name, date: today, amount: 0 },
     // 入力値の検証
     resolver: zodResolver(schema),
   })
@@ -53,34 +69,55 @@ export default function AcountAddForm({
     setLoading(true)
 
     try {
-      // 金融機関登録済みチェック
-      const { data: FinancialInstitution, error: errorFinancialInstitution } =
-        await supabase
-          .from('FinancialInstitution')
-          .select('id')
-          .eq('name', data.financeInstitution)
-          .single()
-      // 登録済みの場合、メッセージ表示
-      if (FinancialInstitution !== null) {
-        setMessage('既に登録されています。')
+      // 金融機関ID取得
+      const { data: FinancialInstitution } = await supabase
+        .from('FinancialInstitution')
+        .select('id')
+        .eq('name', data.financeInstitution)
+        .single()
+      if (FinancialInstitution === null) {
+        setMessage('入力した金融機関が見つかりません。')
         return
       }
-      if (FinancialInstitution === null) {
-        // 金融機関登録
-        const { error: errorInputFinancialInstitution } = await supabase
-          .from('FinancialInstitution')
-          .insert({
-            user_id: id,
-            name: data.financeInstitution,
-            usage: data.usage,
-          })
+      // 金額登録済みチェック
+      const { data: Asset, error: errorAsset } = await supabase
+        .from('Asset')
+        .select('date')
+        .eq('user_id', id)
+        .eq('financial_institution_id', FinancialInstitution.id)
+        .eq('date', data.date)
+        .single()
+      if (Asset !== null) {
+        // 金額更新
+        const { error: errorUpdateAmount } = await supabase
+          .from('Asset')
+          .update({ amount: data.amount })
+          .eq('user_id', id)
+          .eq('financial_institution_id', FinancialInstitution.id)
+          .eq('date', data.date)
+
+        // エラーチェック
+        if (errorUpdateAmount) {
+          setMessage('エラーが発生しました。' + errorUpdateAmount.message)
+          return
+        }
+      } else if (Asset === null) {
+        // 金額登録
+        const { error: errorInputAmount } = await supabase
+          .from('Asset')
+          .insert([
+            {
+              user_id: id,
+              financial_institution_id: FinancialInstitution.id,
+              date: data.date,
+              amount: data.amount,
+            },
+          ])
           .eq('user_id', id)
 
         // エラーチェック
-        if (errorInputFinancialInstitution) {
-          setMessage(
-            'エラーが発生しました。' + errorInputFinancialInstitution.message
-          )
+        if (errorInputAmount) {
+          setMessage('エラーが発生しました。' + errorInputAmount.message)
           return
         }
       }
@@ -100,7 +137,6 @@ export default function AcountAddForm({
     setOpen(false)
     router.refresh()
   }
-
   // 金額入力モーダルウィンドウを閉じる
   if (!open) {
     return null
@@ -122,10 +158,10 @@ export default function AcountAddForm({
               </Button>
             </div>
             <Typography variant='h4' color='blue-gray'>
-              金融機関登録
+              金額入力
             </Typography>
             <Typography color='gray' className='mt-1 font-normal'>
-              Enter your financial institution to register
+              Enter your amounts to register
             </Typography>
             <form
               onSubmit={handleSubmit(onSubmit)}
@@ -135,20 +171,31 @@ export default function AcountAddForm({
                 <Input
                   size='lg'
                   label='金融機関'
-                  type='text'
                   {...register('financeInstitution', { required: true })}
                 />
                 <div className='text-center text-sm text-red-500'>
                   {errors.financeInstitution?.message}
                 </div>
                 <Input
+                  type='date'
                   size='lg'
-                  label='使い方'
-                  type='text'
-                  {...register('usage', { required: true })}
+                  label='日付'
+                  {...register('date', { required: true, valueAsDate: true })}
                 />
                 <div className='text-center text-sm text-red-500'>
-                  {errors.usage?.message}
+                  {errors.date?.message}
+                </div>
+                <Input
+                  type='number'
+                  size='lg'
+                  label='金額 : 円'
+                  {...register('amount', {
+                    required: true,
+                    valueAsNumber: true,
+                  })}
+                />
+                <div className='text-center text-sm text-red-500'>
+                  {errors.amount?.message}
                 </div>
               </div>
               {loading ? (
